+++ docs/CHANGELOG.rst
@@ -14,6 +14,8 @@ Fixed
 -----
 - Make manager more robust to ORM/database failures during data object processing
 - Rebuild the ElasticSearch index after permission is removed from an object
+- Trim ``Data.process_error``, ``Data.process_warning`` and
+  ``Data.process_info`` fields before saving them
 
 
 ==================