+++ docs/CHANGELOG.rst
@@ -19,6 +19,7 @@ Fixed
 - Make sure values in ``Data.process_error``, ``Data.process_warning`` and
   ``Data.process_info`` cannot be overwritten
 - Handle missing ``Data`` objects in ``hydrate_input_references`` function
+- Make executor fail early when executed twice on the same data directory
 
 
 ==================