+++ docs/CHANGELOG.rst
@@ -25,6 +25,8 @@ Added
 -----
 - Add Ubuntu 17.10 base Docker image
 - Add ``descriptor_completed`` field to the ``Entity`` filter
+- Validate manager semaphors after each test case, to ease debugging of tests
+  which execute processes
 
 Fixed
 -----