+++ docs/CHANGELOG.rst
@@ -30,16 +30,25 @@ Added
 
 Fixed
 -----
-- Make manager more robust to ORM/database failures during data object processing
+- Make parallel test suite worker threads clean up after initialization failures
+- Add mechanism to override the manager's control channel prefix from the
+  environment
+
+
+==================
+6.0.1 - 2018-01-29
+==================
+
+Fixed
+-----
+- Make manager more robust to ORM/database failures during data object
+  processing
 - Rebuild the ElasticSearch index after permission is removed from an object
 - Trim ``Data.process_error``, ``Data.process_warning`` and
   ``Data.process_info`` fields before saving them
 - Make sure values in ``Data.process_error``, ``Data.process_warning`` and
   ``Data.process_info`` cannot be overwritten
 - Handle missing ``Data`` objects in ``hydrate_input_references`` function
-- Make parallel test suite worker threads clean up after initialization failures
-- Add mechanism to override the manager's control channel prefix from the
-  environment
 - Make executor fail early when executed twice on the same data directory